updates for the moulibft tests

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: jkellehe <jkellehe@student.42.us.org>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/18 18:11:08 by jkellehe          #+#    #+#             */
-/*   Updated: 2018/06/21 22:31:58 by jkellehe         ###   ########.fr       */
+/*   Updated: 2018/07/06 18:23:28 by jkellehe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s[i] != '\0')
